更新调度配置

+++ app/src/main/assets/powercfg/kona/conservative.sh
@@ -24,33 +24,27 @@ init () {
 }
 if [[ "$action" = "init" ]]; then
   init
-	exit 0
+  exit 0
 fi
-
-stop perfd
-
-echo 0 > /sys/module/msm_thermal/core_control/enabled
-echo 0 > /sys/module/msm_thermal/vdd_restriction/enabled
-echo N > /sys/module/msm_thermal/parameters/enabled
 
 governor0=`cat /sys/devices/system/cpu/cpufreq/policy0/scaling_governor`
 governor4=`cat /sys/devices/system/cpu/cpufreq/policy4/scaling_governor`
 governor7=`cat /sys/devices/system/cpu/cpufreq/policy7/scaling_governor`
 
 if [[ ! "$governor0" = "schedutil" ]]; then
-	echo 'schedutil' > /sys/devices/system/cpu/cpufreq/policy0/scaling_governor
+  echo 'schedutil' > /sys/devices/system/cpu/cpufreq/policy0/scaling_governor
 fi
 if [[ ! "$governor4" = "schedutil" ]]; then
-	echo 'schedutil' > /sys/devices/system/cpu/cpufreq/policy4/scaling_governor
+  echo 'schedutil' > /sys/devices/system/cpu/cpufreq/policy4/scaling_governor
 fi
 if [[ ! "$governor7" = "schedutil" ]]; then
-	echo 'schedutil' > /sys/devices/system/cpu/cpufreq/policy7/scaling_governor
+  echo 'schedutil' > /sys/devices/system/cpu/cpufreq/policy7/scaling_governor
 fi
-
 
 governor_backup () {
   local governor_backup=/cache/governor_backup.prop
-  if [[ ! -f $governor_backup ]]; then
+  local backup_state=`getprop vtools.dev_freq_backup`
+  if [[ ! -f $governor_backup ]] || [[ "$backup_state" != "true" ]]; then
     echo '' > $governor_backup
     local dir=/sys/class/devfreq
     for file in `ls $dir`; do
@@ -59,29 +53,38 @@ governor_backup () {
         echo "$file#$governor" >> $governor_backup
       fi
     done
+    setprop vtools.dev_freq_backup true
   fi
 }
 
 governor_performance () {
   governor_backup
+
   local dir=/sys/class/devfreq
-  for file in `ls $dir`; do
-    if [ -f $dir/$file/governor ]; then
-      echo $dir/$file/governor
-      echo performance > $dir/$file/governor
-    fi
-  done
+  local governor_backup=/cache/governor_backup.prop
+  local backup_state=`getprop vtools.dev_freq_backup`
+
+  if [[ -f "$governor_backup" ]] && [[ "$backup_state" == "true" ]]; then
+    for file in `ls $dir`; do
+      if [ -f $dir/$file/governor ]; then
+        # echo $dir/$file/governor
+        echo performance > $dir/$file/governor
+      fi
+    done
+  fi
 }
 
 governor_restore () {
   local governor_backup=/cache/governor_backup.prop
-  local dir=/sys/class/devfreq
-  if [[ -f "$governor_backup" ]]; then
-      while read line; do
-        if [[ "$line" != "" ]]; then
-            echo ${line#*#} > $dir/${line%#*}/governor
-        fi
-      done < /cache/governor_backup.prop
+  local backup_state=`getprop vtools.dev_freq_backup`
+
+  if [[ -f "$governor_backup" ]] && [[ "$backup_state" == "true" ]]; then
+    local dir=/sys/class/devfreq
+    while read line; do
+      if [[ "$line" != "" ]]; then
+        echo ${line#*#} > $dir/${line%#*}/governor
+      fi
+    done < $governor_backup
   fi
 }
 
@@ -141,13 +144,29 @@ if [[ "$gpu_min_pl" -lt 0 ]];then
 fi;
 
 if [[ ! "$gpu_governor" = "msm-adreno-tz" ]]; then
-	echo 'msm-adreno-tz' > /sys/class/kgsl/kgsl-3d0/devfreq/governor
+  echo 'msm-adreno-tz' > /sys/class/kgsl/kgsl-3d0/devfreq/governor
 fi
 
 echo $gpu_max_freq > /sys/class/kgsl/kgsl-3d0/devfreq/max_freq
 echo $gpu_min_freq > /sys/class/kgsl/kgsl-3d0/devfreq/min_freq
 echo $gpu_min_pl > /sys/class/kgsl/kgsl-3d0/min_pwrlevel
 echo $gpu_max_pl > /sys/class/kgsl/kgsl-3d0/max_pwrlevel
+
+set_gpu_min_freq() {
+  index=$1
+
+  # GPU频率表
+  gpu_freqs=`cat /sys/class/kgsl/kgsl-3d0/devfreq/available_frequencies`
+
+  target_freq=$(echo $gpu_freqs | awk "{print \$${index}}")
+  if [[ "$target_freq" != "" ]]; then
+    echo $target_freq > /sys/class/kgsl/kgsl-3d0/devfreq/min_freq
+  fi
+
+  # gpu_max_freq=`cat /sys/class/kgsl/kgsl-3d0/devfreq/max_freq`
+  # gpu_min_freq=`cat /sys/class/kgsl/kgsl-3d0/devfreq/min_freq`
+  # echo "Frequency: ${gpu_min_freq} ~ ${gpu_max_freq}"
+}
 
 
 set_input_boost_freq() {
@@ -156,19 +175,19 @@ set_input_boost_freq() {
     local c2="$3"
     local ms="$4"
     echo "0:$c0 1:$c0 2:$c0 3:$c0 4:$c1 5:$c1 6:$c1 7:$c2" > /sys/module/cpu_boost/parameters/input_boost_freq
-	echo $ms > /sys/module/cpu_boost/parameters/input_boost_ms
+  echo $ms > /sys/module/cpu_boost/parameters/input_boost_ms
 }
 
 set_cpu_freq()
 {
   echo $1 $2 $3 $4
-	echo "0:$2 1:$2 2:$2 3:$2 4:$4 5:$4 6:$4 7:$6" > /sys/module/msm_performance/parameters/cpu_max_freq
-	echo $1 > /sys/devices/system/cpu/cpufreq/policy0/scaling_min_freq
-	echo $2 > /sys/devices/system/cpu/cpufreq/policy0/scaling_max_freq
-	echo $3 > /sys/devices/system/cpu/cpufreq/policy4/scaling_min_freq
-	echo $4 > /sys/devices/system/cpu/cpufreq/policy4/scaling_max_freq
-	echo $5 > /sys/devices/system/cpu/cpufreq/policy7/scaling_min_freq
-	echo $6 > /sys/devices/system/cpu/cpufreq/policy7/scaling_max_freq
+  echo "0:$2 1:$2 2:$2 3:$2 4:$4 5:$4 6:$4 7:$6" > /sys/module/msm_performance/parameters/cpu_max_freq
+  echo $1 > /sys/devices/system/cpu/cpufreq/policy0/scaling_min_freq
+  echo $2 > /sys/devices/system/cpu/cpufreq/policy0/scaling_max_freq
+  echo $3 > /sys/devices/system/cpu/cpufreq/policy4/scaling_min_freq
+  echo $4 > /sys/devices/system/cpu/cpufreq/policy4/scaling_max_freq
+  echo $5 > /sys/devices/system/cpu/cpufreq/policy7/scaling_min_freq
+  echo $6 > /sys/devices/system/cpu/cpufreq/policy7/scaling_max_freq
 }
 
 sched_config() {
@@ -183,65 +202,75 @@ sched_config() {
   echo "$4" > /proc/sys/kernel/sched_group_upmigrate
 }
 
+sched_limit() {
+  echo $1 > /sys/devices/system/cpu/cpufreq/policy0/schedutil/down_rate_limit_us
+  echo $2 > /sys/devices/system/cpu/cpufreq/policy0/schedutil/up_rate_limit_us
+  echo $3 > /sys/devices/system/cpu/cpufreq/policy4/schedutil/down_rate_limit_us
+  echo $4 > /sys/devices/system/cpu/cpufreq/policy4/schedutil/up_rate_limit_us
+  echo $5 > /sys/devices/system/cpu/cpufreq/policy7/schedutil/down_rate_limit_us
+  echo $6 > /sys/devices/system/cpu/cpufreq/policy7/schedutil/up_rate_limit_us
+}
+
 if [[ "$action" = "powersave" ]]; then
   echo 1 > /sys/devices/system/cpu/cpu4/core_ctl/enable
   echo 1 > /sys/devices/system/cpu/cpu7/core_ctl/enable
   echo 1 > /sys/devices/system/cpu/cpu4/core_ctl/min_cpus
 
-	set_cpu_freq 300000 1612800 710400 1286400 844800 1305600
+  set_cpu_freq 300000 1804800 710400 1478400 844800 1632000
   set_input_boost_freq 1248000 0 0 40
 
-	echo $gpu_min_pl > /sys/class/kgsl/kgsl-3d0/default_pwrlevel
-	echo 0 > /proc/sys/kernel/sched_boost
+  echo $gpu_min_pl > /sys/class/kgsl/kgsl-3d0/min_pwrlevel
+  echo 0 > /proc/sys/kernel/sched_boost
 
-	echo 1248000 > /sys/devices/system/cpu/cpufreq/policy0/schedutil/hispeed_freq
-	echo 710400 > /sys/devices/system/cpu/cpufreq/policy4/schedutil/hispeed_freq
-	echo 844800 > /sys/devices/system/cpu/cpufreq/policy7/schedutil/hispeed_freq
+  echo 1248000 > /sys/devices/system/cpu/cpufreq/policy0/schedutil/hispeed_freq
+  echo 710400 > /sys/devices/system/cpu/cpufreq/policy4/schedutil/hispeed_freq
+  echo 844800 > /sys/devices/system/cpu/cpufreq/policy7/schedutil/hispeed_freq
 
   echo 0-2 > /dev/cpuset/background/cpus
   echo 0-3 > /dev/cpuset/system-background/cpus
 
-	sched_config "85 85" "96 96" "160" "260"
-  echo 0 > /sys/devices/system/cpu/cpu7/online
+  sched_config "85 85" "96 96" "160" "260"
 
-	exit 0
+  sched_limit 0 0 0 5000 0 5000
+
+  exit 0
 fi
 
 if [[ "$action" = "balance" ]]; then
-  echo 1 > /sys/devices/system/cpu/cpu7/online
   echo 1 > /sys/devices/system/cpu/cpu4/core_ctl/enable
   echo 1 > /sys/devices/system/cpu/cpu7/core_ctl/enable
   echo 2 > /sys/devices/system/cpu/cpu4/core_ctl/min_cpus
 
-	set_cpu_freq 300000 1708800 710400 1862400 844800 2073600
+  set_cpu_freq 300000 1804800 710400 1862400 844800 2073600
   set_input_boost_freq 1478400 0 0 40
 
-	echo $gpu_min_pl > /sys/class/kgsl/kgsl-3d0/default_pwrlevel
-	echo 0 > /proc/sys/kernel/sched_boost
+  echo $gpu_min_pl > /sys/class/kgsl/kgsl-3d0/min_pwrlevel
+  echo 0 > /proc/sys/kernel/sched_boost
 
-	echo 1420800 > /sys/devices/system/cpu/cpufreq/policy0/schedutil/hispeed_freq
-	echo 1056000 > /sys/devices/system/cpu/cpufreq/policy4/schedutil/hispeed_freq
-	echo 1305600 > /sys/devices/system/cpu/cpufreq/policy7/schedutil/hispeed_freq
+  echo 1420800 > /sys/devices/system/cpu/cpufreq/policy0/schedutil/hispeed_freq
+  echo 1056000 > /sys/devices/system/cpu/cpufreq/policy4/schedutil/hispeed_freq
+  echo 1305600 > /sys/devices/system/cpu/cpufreq/policy7/schedutil/hispeed_freq
 
   echo 0-2 > /dev/cpuset/background/cpus
   echo 0-3 > /dev/cpuset/system-background/cpus
 
-	sched_config "78 85" "89 96" "120" "200"
+  sched_config "78 85" "89 96" "120" "200"
 
-	exit 0
+  sched_limit 0 0 0 500 0 500
+
+  exit 0
 fi
 
 if [[ "$action" = "performance" ]]; then
-  echo 1 > /sys/devices/system/cpu/cpu7/online
   echo 3 > /sys/devices/system/cpu/cpu4/core_ctl/min_cpus
   echo 0 > /sys/devices/system/cpu/cpu4/core_ctl/enable
   echo 0 > /sys/devices/system/cpu/cpu7/core_ctl/enable
 
-	set_cpu_freq 300000 1804800 710400 2419200 825600 2841600
+  set_cpu_freq 300000 1804800 710400 2419200 825600 2841600
   set_input_boost_freq 1420800 1286400 1305600 40
 
-	echo `expr $gpu_min_pl - 1` > /sys/class/kgsl/kgsl-3d0/default_pwrlevel
-	echo 0 > /proc/sys/kernel/sched_boost
+  echo `expr $gpu_min_pl - 1` > /sys/class/kgsl/kgsl-3d0/min_pwrlevel
+  echo 0 > /proc/sys/kernel/sched_boost
 
   echo 1612800 > /sys/devices/system/cpu/cpufreq/policy0/schedutil/hispeed_freq
   echo 1766400 > /sys/devices/system/cpu/cpufreq/policy4/schedutil/hispeed_freq
@@ -250,31 +279,34 @@ if [[ "$action" = "performance" ]]; then
   echo 0-1 > /dev/cpuset/background/cpus
   echo 0-3 > /dev/cpuset/system-background/cpus
 
-	sched_config "65 78" "75 88" "85" "100"
+  sched_config "65 78" "75 88" "85" "100"
 
-	exit 0
+  sched_limit 0 0 0 0 0 0
+
+  exit 0
 fi
 
 if [[ "$action" = "fast" ]]; then
-  echo 1 > /sys/devices/system/cpu/cpu7/online
   echo 3 > /sys/devices/system/cpu/cpu4/core_ctl/min_cpus
   echo 0 > /sys/devices/system/cpu/cpu4/core_ctl/enable
   echo 0 > /sys/devices/system/cpu/cpu7/core_ctl/enable
 
-	set_cpu_freq 1075200 1804800 1382400 2600000 1305600 3200000
+  set_cpu_freq 1248000 1804800 1478400 2600000 1516800 3200000
   set_input_boost_freq 1804800 1670400 1862400 80
 
-	echo 1612800 > /sys/devices/system/cpu/cpufreq/policy0/schedutil/hispeed_freq
-	echo 1670400 > /sys/devices/system/cpu/cpufreq/policy4/schedutil/hispeed_freq
-	echo 1862400 > /sys/devices/system/cpu/cpufreq/policy7/schedutil/hispeed_freq
+  echo 1612800 > /sys/devices/system/cpu/cpufreq/policy0/schedutil/hispeed_freq
+  echo 1670400 > /sys/devices/system/cpu/cpufreq/policy4/schedutil/hispeed_freq
+  echo 1862400 > /sys/devices/system/cpu/cpufreq/policy7/schedutil/hispeed_freq
 
-	echo `expr $gpu_min_pl - 2` > /sys/class/kgsl/kgsl-3d0/default_pwrlevel
-	echo 1 > /proc/sys/kernel/sched_boost
+  echo `expr $gpu_min_pl - 2` > /sys/class/kgsl/kgsl-3d0/min_pwrlevel
+  echo 1 > /proc/sys/kernel/sched_boost
 
   echo 0 > /dev/cpuset/background/cpus
   echo 0-3 > /dev/cpuset/system-background/cpus
 
-	sched_config "62 78" "72 85" "85" "100"
+  sched_config "62 78" "72 85" "85" "100"
 
-	exit 0
+  sched_limit 5000 0 2000 0 2000 0
+
+  exit 0
 fi
